+++ src/test/java/com/global/api/tests/network/nts/NtsRequestToBalanceTest.java
@@ -25,6 +25,7 @@ import com.global.api.tests.testdata.NtsTestCards;
 import org.junit.Test;
 
 import java.math.BigDecimal;
+import java.util.ArrayList;
 import java.util.LinkedList;
 
 import static org.junit.Assert.*;
@@ -592,12 +593,26 @@ public class NtsRequestToBalanceTest {
         BatchSummary summary = batchSummary.resubmitTransactions(list);
         assertNotNull(summary);
 
-        BatchSummary newSummary = batchSummary.resubmitTransactions(false,summary.getNonApprovedDataCollectToken());
-        assertNotNull(newSummary);
-
-        NetworkService.resubmitBatchClose(batchClose.getTransactionToken())
-                .withForceToHost(true)
-                .execute();
+        BatchSummary newSummary1 = batchSummary.resubmitTransactions(false,summary.getNonApprovedDataCollectToken(),false);
+        assertNotNull(newSummary1);
+        BatchSummary newSummary2 = batchSummary.resubmitTransactions(false,newSummary1.getNonApprovedDataCollectToken(),false);
+        assertNotNull(newSummary2);
+        BatchSummary newSummary3 = batchSummary.resubmitTransactions(false,newSummary2.getNonApprovedDataCollectToken(),false);
+        assertNotNull(newSummary3);
+        BatchSummary newSummary4 = batchSummary.resubmitTransactions(false,newSummary3.getNonApprovedDataCollectToken(),false);
+        assertNotNull(newSummary4);
+        BatchSummary newSummary5 = batchSummary.resubmitTransactions(false,newSummary4.getNonApprovedDataCollectToken(),false);
+        assertNotNull(newSummary5);
+        BatchSummary newSummary6 = batchSummary.resubmitTransactions(false,newSummary5.getNonApprovedDataCollectToken(),false);
+        assertNotNull(newSummary6);
+        BatchSummary newSummary7 = batchSummary.resubmitTransactions(false,newSummary6.getNonApprovedDataCollectToken(),false);
+        assertNotNull(newSummary7);
+        BatchSummary newSummary8 = batchSummary.resubmitTransactions(false,newSummary7.getNonApprovedDataCollectToken(),false);
+        assertNotNull(newSummary8);
+        BatchSummary newSummary9 = batchSummary.resubmitTransactions(false,newSummary8.getNonApprovedDataCollectToken(),false);
+        assertNotNull(newSummary9);
+        BatchSummary newSummary10 = batchSummary.resubmitTransactions(false,newSummary9.getNonApprovedDataCollectToken(),false);
+        assertNotNull(newSummary10);
 
     }
 
@@ -634,11 +649,11 @@ public class NtsRequestToBalanceTest {
         BatchSummary summary = batchSummary.resubmitTransactions(list);
         assertNotNull(summary);
 
-        BatchSummary newSummary = batchSummary.resubmitTransactions(true,summary.getFormatErrorDataCollectToken());
+        BatchSummary newSummary = batchSummary.resubmitTransactions(true,summary.getFormatErrorDataCollectToken(),false);
         assertNotNull(newSummary);
 
         batchSummary.setHostResponseCode("79");
-        BatchSummary newSummary1 = batchSummary.resubmitTransactions(true,newSummary.getFormatErrorDataCollectToken());
+        BatchSummary newSummary1 = batchSummary.resubmitTransactions(true,newSummary.getFormatErrorDataCollectToken(),false);
         assertNotNull(newSummary1);
     }
     @Test
@@ -676,6 +691,101 @@ public class NtsRequestToBalanceTest {
                 .withForceToHost(true)
                 .execute();
         assertNotNull(forceBatchCloseResponse);
+
+    }
+
+    @Test
+    public void test_BatchCloseIssue_10213_RTB_01() throws ApiException {
+
+        Transaction t= creditSale(10.11);
+        Transaction t1 =creditSale(20.22);
+
+        NtsRequestMessageHeader ntsRequestMessageHeader = new NtsRequestMessageHeader();
+        ntsRequestMessageHeader.setTerminalDestinationTag("478");
+        ntsRequestMessageHeader.setPinIndicator(PinIndicator.WithPin);
+        ntsRequestMessageHeader.setNtsMessageCode(NtsMessageCode.RequestToBalacnce);
+        ntsRequestMessageHeader.setPinIndicator(PinIndicator.NotPromptedPin);
+
+        priorMessageInformation =new PriorMessageInformation();
+        priorMessageInformation.setResponseTime("999");
+        priorMessageInformation.setConnectTime("999");
+        priorMessageInformation.setMessageReasonCode("01");
+        ntsRequestMessageHeader.setPriorMessageInformation(priorMessageInformation);
+
+        NtsRequestToBalanceData data = new NtsRequestToBalanceData(batchProvider.getSequenceNumber(), new BigDecimal(1), "Version");
+        Transaction batchClose = BatchService.closeBatch(BatchCloseType.EndOfShift,
+                        ntsRequestMessageHeader, batchProvider.getBatchNumber(), 2
+                        , new BigDecimal(30.33), BigDecimal.ONE, data)
+                .execute();
+        assertNotNull(batchClose);
+
+        BatchSummary batchSummary = batchClose.getBatchSummary();
+        LinkedList<String> list = new LinkedList<>();
+        list.add(t.getTransactionToken());
+        list.add(t1.getTransactionToken());
+
+        BatchSummary summary = batchSummary.resubmitTransactions(list);
+        assertNotNull(summary);
+
+        ArrayList<String> tokens = new ArrayList<>(summary.getNonApprovedDataCollectToken());
+        for(String token :tokens){
+            for(int i = 0; i < 10; i++){
+                Transaction dataCollect = NetworkService.resubmitDataCollect(token)
+                        .execute();
+            }
+        }
+
+       ArrayList<String> list1 = new ArrayList<>();
+        if(summary.getNonApprovedDataCollectToken() != null){
+            list1.addAll(summary.getNonApprovedDataCollectToken());
+        }
+        if(summary.getAllDataCollectToken() != null){
+            list1.addAll(summary.getAllDataCollectToken());
+        }
+
+        BatchSummary newSummary = batchSummary.resubmitTransactions(list1);
+        assertNotNull(newSummary);
+
+    }
+
+    @Test
+    public void test_BatchCloseIssue_10213_RTB_40() throws ApiException {
+
+        Transaction t= creditSale(10.11);
+        Transaction t1 =creditSale(20.22);
+
+        NtsRequestMessageHeader ntsRequestMessageHeader = new NtsRequestMessageHeader();
+        ntsRequestMessageHeader.setTerminalDestinationTag("478");
+        ntsRequestMessageHeader.setPinIndicator(PinIndicator.WithPin);
+        ntsRequestMessageHeader.setNtsMessageCode(NtsMessageCode.RequestToBalacnce);
+        ntsRequestMessageHeader.setPinIndicator(PinIndicator.NotPromptedPin);
+
+        priorMessageInformation =new PriorMessageInformation();
+        priorMessageInformation.setResponseTime("999");
+        priorMessageInformation.setConnectTime("999");
+        priorMessageInformation.setMessageReasonCode("01");
+        ntsRequestMessageHeader.setPriorMessageInformation(priorMessageInformation);
+
+        NtsRequestToBalanceData data = new NtsRequestToBalanceData(batchProvider.getSequenceNumber(), new BigDecimal(1), "Version");
+        Transaction batchClose = BatchService.closeBatch(BatchCloseType.EndOfShift,
+                        ntsRequestMessageHeader, batchProvider.getBatchNumber(), 2
+                        , new BigDecimal(30.33), BigDecimal.ONE, data)
+                .execute();
+        assertNotNull(batchClose);
+
+        BatchSummary batchSummary = batchClose.getBatchSummary();
+        LinkedList<String> list = new LinkedList<>();
+        list.add(t.getTransactionToken());
+        list.add(t1.getTransactionToken());
+
+        BatchSummary summary = batchSummary.resubmitTransactions(list);
+        assertNotNull(summary);
+
+        for(int i = 0; i < 10; i++){
+            NetworkService.resubmitBatchClose(summary.getTransactionToken())
+                    .execute();
+        }
+
 
     }
 }